Develop

+++ appcenter-pre-build.sh
@@ -1,5 +1,15 @@
 #!/usr/bin/env bash
 
-npx jetify
+if [ "$PLATFORM" == "ios" ] ; then
+    cd ./ios && rm -rf Pods && pod cache clean --all && pod install && cd ..
+else
+    npx jetify
+fi
 
-cd ./ios && rm -rf Pods && pod cache clean --all && pod install && cd ..
+# Creates an .env from ENV variables for use with react-native-config
+ENV_WHITELIST=${ENV_WHITELIST:-"^RN_"}
+printf "Creating an .env file with the following whitelist:\n"
+printf "%s\n" $ENV_WHITELIST
+set | egrep -e $ENV_WHITELIST | sed 's/^RN_//g' > .env
+printf "\n.env created with contents:\n\n"
+cat .env
